+++ packages/web3-utils/types/tests/sha3-raw-test.ts
@@ -25,7 +25,7 @@ import {sha3Raw} from 'web3-utils';
 
 // $ExpectType string
 sha3Raw('234');
-// $ExpectType string | null
+// $ExpectType string
 sha3Raw(Buffer.from('123'));
 // $ExpectType string
 sha3Raw(new BN(3));