+++ CHANGELOG.md
@@ -586,6 +586,7 @@ Released with 1.0.0-beta.37 code base.
 ## [1.7.6]
 
 ### Changed
+-  Updated `sha3` and `sha3Raw` type definition to accept `Buffer`
 
 ### Fixed
 -  Browser builds support polyfills (#5031) (#5053) (#4659) (#4767)