+++ CHANGELOG.md
@@ -583,6 +583,8 @@ Released with 1.0.0-beta.37 code base.
 ### Security
 -  Updated `got` lib version and fixed other libs using npm audit fix (#5178) (#5254)
 
+## [Unreleased]
+
 ## [1.7.6]
 
 ### Changed