+++ CHANGELOG.md
@@ -591,7 +591,7 @@ Released with 1.0.0-beta.37 code base.
 -   Fix typos in web3-eth-accounts.rst & TESTING.md (#5047)
 -   Fix remove wallet using an index when an account address and address lowercase are equal (#5049)
 -   Improve README.md & Fix typos (#4848)
--   Add optional hex formatting parameter for getTransactionrReceipt (#5153)
+-   Add optional hex formatting parameter for getTransactionReceipt (#5153)
 -   Fix transactionRoot -> transactionsRoot in BlockHeader (#5083)
 -   Fix Promise in Accounts.signTransaction() throwing errors that cannot be caught (#4724)
 -   Fixed unit tests & removed dead code for web3-providers-http (#5228)
@@ -602,7 +602,7 @@ Released with 1.0.0-beta.37 code base.
 
 ## [Unreleased]
 
-## [4.0.0-alpha.1]
+## [4.0.0-alpha.0]
 
 ### Added